+++ apps/hardware_benchmarks/tests/bitwise/bitwise_generator.cpp
@@ -6,8 +6,8 @@ using namespace Halide;
 
 class UnitTestBitwise : public Halide::Generator<UnitTestBitwise> {
 public:
-    Input<Buffer<int16_t>>  input{"input", 2};
-    Output<Buffer<int16_t>> output{"output", 2};
+    Input<Buffer<uint8_t>>  input{"input", 2};
+    Output<Buffer<uint8_t>> output{"output", 2};
 
     void generate() {
         /* THE ALGORITHM */
@@ -24,7 +24,7 @@ public:
         bitwise_xor(x,y) = bitwise_inv(x,y) ^ bitwise_or(x,y);
 
         Func hw_output("hw_output");
-        hw_output(x, y) = cast<int16_t>(bitwise_xor(x, y));
+        hw_output(x, y) = cast<uint8_t>(bitwise_xor(x, y));
         output(x, y) = hw_output(x,y);
 
         /* THE SCHEDULE */
