+++ apps/hardware_benchmarks/tests/conv_1_2/process.cpp
@@ -11,7 +11,7 @@ using namespace Halide::Runtime;
 
 int main(int argc, char **argv) {
 
-  OneInOneOut_ProcessController<uint16_t> processor("conv_1_2",
+  OneInOneOut_ProcessController<uint8_t> processor("conv_1_2",
                                           {
                                             {"cpu",
                                                 [&]() { conv_1_2(processor.input, processor.output); }
@@ -22,8 +22,8 @@ int main(int argc, char **argv) {
                                             }
                                           });
 
-  processor.input = Buffer<uint16_t>(64, 64);
-  processor.output = Buffer<uint16_t>(63, 64);
+  processor.input = Buffer<uint8_t>(64, 64);
+  processor.output = Buffer<uint8_t>(63, 64);
   
   processor.process_command(argc, argv);
