+++ apps/hardware_benchmarks/tests/conv_3_3/process.cpp
@@ -11,7 +11,7 @@ using namespace Halide::Runtime;
 
 int main(int argc, char **argv) {
 
-  OneInOneOut_ProcessController<uint16_t> processor("conv_3_3",
+  OneInOneOut_ProcessController<uint8_t> processor("conv_3_3",
                                             {
                                               {"cpu",
                                                   [&]() { conv_3_3(processor.input, processor.output); }
@@ -23,8 +23,8 @@ int main(int argc, char **argv) {
 
                                             });
 
-  processor.input = Buffer<uint16_t>(64, 64);
-  processor.output = Buffer<uint16_t>(62, 62);
+  processor.input = Buffer<uint8_t>(64, 64);
+  processor.output = Buffer<uint8_t>(62, 62);
   
   processor.process_command(argc, argv);
