+++ apps/hardware_benchmarks/tests/sminmax/process.cpp
@@ -11,7 +11,7 @@ using namespace Halide::Runtime;
 
 int main(int argc, char **argv) {
 
-  OneInOneOut_ProcessController<int16_t> processor("sminmax",
+  OneInOneOut_ProcessController<uint8_t> processor("sminmax",
                                           {
                                             {"cpu",
                                                 [&]() { sminmax(processor.input, processor.output); }
@@ -22,8 +22,8 @@ int main(int argc, char **argv) {
                                             }
                                           });
 
-  processor.input = Buffer<int16_t>(64, 64);
-  processor.output = Buffer<int16_t>(64, 64);
+  processor.input = Buffer<uint8_t>(64, 64);
+  processor.output = Buffer<uint8_t>(64, 64);
   
   processor.process_command(argc, argv);
