+++ apps/hardware_benchmarks/tests/sminmax/sminmax_generator.cpp
@@ -6,8 +6,8 @@ using namespace Halide;
 
 class UnitTestSminmax : public Halide::Generator<UnitTestSminmax> {
 public:
-    Input<Buffer<int16_t>>  input{"input", 2};
-    Output<Buffer<int16_t>> output{"output", 2};
+    Input<Buffer<uint8_t>>  input{"input", 2};
+    Output<Buffer<uint8_t>> output{"output", 2};
 
     void generate() {
         /* THE ALGORITHM */
@@ -23,7 +23,7 @@ public:
         sclamp(x,y) = clamp( hw_input(x,y), 1, 150 );
 
         Func hw_output("hw_output");
-        hw_output(x, y) = cast<int16_t>(smin(x,y) + smax(x,y) - sclamp(x,y));
+        hw_output(x, y) = cast<uint8_t>(smin(x,y) + smax(x,y) - sclamp(x,y));
         output(x, y) = hw_output(x,y);
 
         /* THE SCHEDULE */
